Changed chunks from HashMap to Vec<> Performance drastically decreased, because pixels are drawn 1 pixel per drawcall
Now because the chunk is not a sparse grid
but fully filled with empty pixels
Every pixel is being redrawn every frame

Fix this by first drawing the pixel colours to a texture
Then drawing that texture

+++ src/main.rs
@@ -19,7 +19,7 @@ pub fn window_settings() -> Conf {
     }
 }
 
-const CHUNK_SIZE: (i32, i32) = (160, 90);
+const CHUNK_SIZE: (usize, usize) = (160, 90);
 const RENDER_SIZE: (u32, u32) = (240, 125);
 
 #[main(window_settings)]
